+++ __tests__/exchanges/binanceje-client.spec.js
@@ -67,6 +67,7 @@ testClient({
 
   hasTickers: true,
   hasTrades: true,
+  hasCandles: true,
   hasLevel2Snapshots: true,
   hasLevel2Updates: true,
   hasLevel3Snapshots: false,
@@ -91,6 +92,8 @@ testClient({
   trade: {
     hasTradeId: true,
   },
+
+  candle: {},
 
   l2snapshot: {
     hasTimestampMs: false,
