+++ __tests__/exchanges/binanceus-client.spec.js
@@ -42,6 +42,7 @@ testClient({
 
   hasTickers: true,
   hasTrades: true,
+  hasCandles: true,
   hasLevel2Snapshots: true,
   hasLevel2Updates: true,
   hasLevel3Snapshots: false,
@@ -66,6 +67,8 @@ testClient({
   trade: {
     hasTradeId: true,
   },
+
+  candle: {},
 
   l2snapshot: {
     hasTimestampMs: false,
