+++ src/exchanges/binance-client.js
@@ -11,41 +11,7 @@ const Level2Snapshot = require("../level2-snapshot");
 const SmartWss = require("../smart-wss");
 const Watcher = require("../watcher");
 const { CandlePeriod } = require("../enums");
-
-function candlePeriod(p) {
-  switch (p) {
-    case CandlePeriod._1m:
-      return "1m";
-    case CandlePeriod._3m:
-      return "3m";
-    case CandlePeriod._5m:
-      return "5m";
-    case CandlePeriod._15m:
-      return "15m";
-    case CandlePeriod._30m:
-      return "30m";
-    case CandlePeriod._1h:
-      return "1h";
-    case CandlePeriod._2h:
-      return "2h";
-    case CandlePeriod._4h:
-      return "4h";
-    case CandlePeriod._6h:
-      return "6h";
-    case CandlePeriod._8h:
-      return "8h";
-    case CandlePeriod._12h:
-      return "12h";
-    case CandlePeriod._1d:
-      return "1d";
-    case CandlePeriod._3d:
-      return "3d";
-    case CandlePeriod._1w:
-      return "1w";
-    case CandlePeriod._1M:
-      return "1M";
-  }
-}
+const { candlePeriod } = require("./binance-base");
 
 class BinanceClient extends EventEmitter {
   constructor({ useAggTrades = true, requestSnapshot = true, reconnectIntervalMs = 300000 } = {}) {
